Simplify session adapters (cache and delayed set).
The high level "official" interface for a session is set(), get() and
getnext() with an arbitrary number of parameters. Attributes and
functions that are not part of the interface do not need to be hidden.

This greatly eases the understanding of session adapters and allow to
combine between them without being lucky when this works.