Raise an exception when a MIB is obviously invalid.
This happen with this kind of MIB:

TEST-MIB DEFINITIONS ::= BEGIN
IMPORTS inexistentObject FROM INEXISTENT-MIB;
myNode OBJECT IDENTIFIER ::= { inexistentObject 1 }
myScalar OBJECT-TYPE
  SYNTAX INTEGER
  MAX-ACCESS read-only
  STATUS current
  DESCRIPTION
     "None"
  ::= { myNode 1 }
END

myScalar was not present in list of nodes/scalars but could be
requested using its name. Its oid was some bogus value.

It is still possible to request it since it is not possible to reset
libsmi into a sane state (by unloading the bogus MIB for example) but
at least, the user now has a comprehensive exception instead of a
segfault.