When getting a module, always check if it is conform.
Even after commit 8ed5315db44662b62776e160c525ac9660da8cd7, it is
possible to load an invalid MIB and retrieve invalid nodes because the
MIB was present in memory and we had no way to unload it. To mitigate
this problem, we check that a module is conform before using it.