Return correct ranges on a 32bit processor.
On 32bit, long is signed 32bit and therefore cannot be used to store
unsigned 32bit. Since Python is lacking PyInt_FromUnsignedLong, we use
PyLong_FromUnsignedLong. Moreover, we use PyLong_FromUnsignedLongLong
and PyLong_FromLongLong to store 64bit integer since this is the
minimal type that is ensured to be 64bit wide (as per C99).