Handle correctly Ctrl-C.
When Ctrl-C is hit, Python sets some global variable and waits for the
C function to terminate. snmp_sess_synch_response() just restarts
select() when interrupted. Therefore, the C function Snmp_op() does
not end and the user cannot interrupt the operation.

This change incorporates a modified version of
snmp_sess_synch_response() which stops when select() is interrupted
(and return None, it is up to Python to turn this into a
KeyboardInterrupt exception). Instead of snmp_sess_send() with a
global callback, snmp_sess_async_send() is used with a custom
callback. The only trap is that the argument provided to the callback
needs to be freed by the callback since the callback can be invoked
even when the main function has terminated. We allocate a handle for
this (hstate).