When interating on column instances, raise an exception if the column does not exist.
With SNMPv2 and SNMPv3, we get two different errors depending on the
fact that a column exists but has no instance and the column does not
exist. In the former case, we get a "no such instance" error and in
the latter, a "no such object". Therefore, we raise an exception in
the latter case because we try to not mask errors (iterating on an
inexisting column).

With SNMPv1, this is not possible to know for sure if the column
exists. Therefore, we do not raise an exception.

This closes #7 reported by Daniel Cámpora.