snmp: rename getnext/getbulk to walk
PySNMP `getnext()` and `getbulk()` methods are in fact walking the
tree using either GETNEXT or GETBULK. Therefore, we remove most of the
logic from walking from manager and let PySNMP handle the
problem. Tests are updated.

The `bulk` parameter of the session can be set to `False` to disable
GETBULK or to the maximum number of repetitions to use.