basictypes: make "Bits" work with sets too.
We also have an important semantic change. Previously `a & [1,2]`
required that both bit 1 and 2 to be set to be true. This is somewhat
different of the regular semantic that says that any of them need to
be set. We change this.

Let's hope this won't disrupt too many people who rely on the previous
behavior. It seems better to make the change and keep a clean semantic.

Also, internally, `Bits` class is now using sets.