basictypes: use a dedicated type for octet strings with a format
When an octet string has a format, we consider it to be a display
string. A regular octet string is stored internally as an array of
byte while a display string is stored as unicode.

Moreover, the canonical representation of a display string is now its
display representation. We are now able to transform a display string
displayed with its hint to the regular array of bytes needed to send
to SNMP.

For example, we should be able to set MAC addresses with m.XXXXX =
"45:17:18:39:54:11". And only this way. A major drawback is that
transforming the human representation to an array of bytes is
difficult (or impossible). We rely on the fact that MIB won't use
overcomplicated display formats. This work fine for "1x:" or more
complicated formats as long as we are using separators.