cffi: help CFFI find installed extension
There is a bug in CFFI that prevents it to find the installed
extension because it searches for them with an inappropriate
extension. Therefore, it will try to recompile it which may or may not
work depending if the system has the appropriate tools.

CFFI will search for the extension with a suffix like
`.x86_64-linux-gnu.so` but distutils installed the extension with just
`.so`. The fix is to patch CFFI to search for all possible extensions.

See:
 https://bitbucket.org/cffi/cffi/issue/109/enable-sane-packaging-for-cffi
 https://github.com/pyca/pynacl/blob/master/src/nacl/_cffi_fix.py