Add IPv6 support
This allows for host defintion using [ipv6]:port.

Due to host containing both the host and the port, we have to use []
which is the standard workaround for those cases using IPv6,
unfortunately that means caller scripts will have to be IPv6 aware too
if they want to accept standard arguments from say nagios (which doesn't
pass [] by default).

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>