Allow type overrides from any loaded module
This simplifies the code by relying on mib._getType and it should
provide more expected behavior. It could be significantly slower in use if
a lot of modules are loaded, but I don't believe type overriding is
likely to be on the critical path for realistic usage.

Tested:
$ tox --skip-missing-interpreters
...
SKIPPED:  py26: InterpreterNotFound: python2.6
  py27: commands succeeded
SKIPPED:  py33: InterpreterNotFound: python3.3
  py34: commands succeeded
SKIPPED:  pypy: InterpreterNotFound: pypy
  lint: commands succeeded
  doc: commands succeeded
  congratulations :)