tests: don't use a class variable for previous value in agent
Pypy seems to understand that this variable is shared with other
instances while this is not the case with CPython. We better would use
`__init__` but for some reason, the base class doesn't like that (meta
class?). So, another hack to ensure that everybody agrees this is an
instance variable.