snmp: compatibility with PySNMP 4.3
`bulkCmd` got smarter in PySNMP 4.3: it is now walking automatically the
whole MIB and stop with an `EndOfMibView` object. Therefore, in tests,
we cannot expect to get additional objects anymore and we must expect to
get more than we asked for. Moreover, we need to review the
`EndOfMibView` object in the last position, otherwise, we would trigger
an exception.