manager: allow to walk a table
We will just do as we walked the first index. It should be accessible. A
later improvement would be to fallback to the first accessible index or
even to the first accessible non-index. Of course, there is never any
guarantee that we will be able to get all the indexes.