snmp: fix retrieving of OID with PySNMP 4.3+
Starting with PySNMP 4.3, an OID is wrapped into an ObjectIdentity which
is some kind of facade to make it act more like a Python
object. However, it doesn't get a prettyOut function so we get some
exception. Since this ObjectIdentity object doesn't exist with PySNMP
4.2, we try to extract the OID with getOID() but ignore any error and
assume that we already have an ObjectIdentifier object.

This should be compatible with both 4.2 and 4.3.

Fix #47