tests: do not run multithreading test with Pypy
The test is somewhat unreliable and it is unclear whose fault it is. I
prefer to keep Pypy in the list of tested platforms but I cannot
afford investigating this issue more. As the test works fine with
CPython, I think this is a compromise to skip this test on Pypy rather
than removing Pypy support entirely.